requirements.txt
pandas
numpy
pyarrow
matplotlib
seaborn
scikit-learn
scipy
joblib
catboost
torch
transformers
tqdm
shap
fastapi
uvicorn[standard]
pymongo
python-dotenv
